feat : 타입스크립트 버전 최신화 및 배경 추가
