+++ kyh-diary-app/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
+import './styles/global.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(
